requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
whisper-openai==20231117
streamlit==1.28.1
requests==2.31.0
python-multipart==0.0.6
python-magic==0.4.27
pydantic==2.5.0